refactor(convert): extract positioned module
Phase 1 of fulgur-odgo. Move absolute-positioning helpers from
convert/mod.rs into convert/positioned.rs:

- pub(super): AbsCb struct, collect_positioned_children,
  is_absolutely_positioned, build_absolute_pseudo_children,
  try_build_absolute_pseudo_image
- private: is_position_fixed, is_position_static, cb_padding_box,
  resolve_cb_for_absolute, resolve_inset_px,
  build_absolute_pseudo_child

pseudo.rs gains an explicit `use super::positioned::{...}` for the
four items it depends on; this surfaces the cross-module dependency
at the top of the file rather than relying on the parent glob.

No behavioral change. VRT byte-identical against committed goldens.
